pyproject.toml
[project]
name = "gh-repo-summerizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
ghsum = "ghsum.cli:main"
